requirements.txt
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
decorator==4.4.2
feedparser==5.2.1
future==0.18.2
idna==2.9
pycodestyle==2.5.0
pycparser==2.20
python-telegram-bot==12.5.1
requests==2.23.0
requests-cache==0.5.2
six==1.14.0
tornado==6.0.4
urllib3==1.25.8
